feat: upgrade ejkernel to v0.0.2 and optimize RoPE/dataset handling
  This commit includes several improvements across the codebase:

  - Upgrade ejkernel dependency from 0.0.1 to 0.0.2
  - Optimize RoPE embeddings by consolidating dtype conversions to bfloat16
    at frequency generation time rather than after each application
  - Add dataset_split_name parameter support for more flexible dataset loading
    across TextDatasetInform, VisualDatasetInform, and FastDataManager
  - Add save_interval_minutes configuration option to BaseTrainerCfg for
    time-based checkpoint saving
  - Implement automatic data iterator reset in HFDataSource to handle
    exhausted iterators gracefully (controlled by HFDATASOURCE_NONSTOP env var)
  - Switch default GPU attention mechanism from FLASH_ATTN2 to SDPA for
    improved compatibility for old gpus.

+++ easydel/infra/base_config.py
@@ -1259,7 +1259,7 @@ class EasyDeLBaseConfig(PretrainedConfig):
             max_position=self.granted_freq_max_position_embedding,
             base=base or self.rope_theta,
             rope_scaling=rope_config.to_dict(),
-        )
+        ).astype(jnp.bfloat16)
 
         return ModuleCaches(jax.device_put(frequencies, Ns(self.mesh, Ps())))
